organize + init mani code actions
