+++ src/services/manifest/mani-hovers.ts
