+++ src/services/manifest/mani-suggestions.ts
